[CH-11273] Add decimal parsing to heatmap threshold filters (#705)
# Description
- Thankfully a pretty simple fix... `parseFloat` can be transparently substituted for `parseInt` to handle decimal values.

# Notes
- I did a quick audit through the rest of the heatmap file and didn't see other cases where `parseFloat` should be substituted.
- The report page heatmap apparently doesn't have this defect b/c the filters are implemented pretty differently.

# Tests
- Tested locally.
- QA: Decimal values should all work logically in the heatmap threshold filters now.

#### BEFORE: Not respecting the cutoff of 3.57 (3.9 raised to 4.0)

![BEFORE](https://user-images.githubusercontent.com/5652739/108927011-572d7900-75f4-11eb-9d2d-2fe8c8880e05.gif)

#### AFTER: Respecting the cutoff of 3.57 (3.5 raised to 3.6)

![AFTER](https://user-images.githubusercontent.com/5652739/108927016-5ac10000-75f4-11eb-9a47-bd6ca2177081.gif)

+++ app/assets/src/components/views/compare/SamplesHeatmapView/SamplesHeatmapView.jsx
@@ -985,11 +985,11 @@ class SamplesHeatmapView extends React.Component {
           return false;
         }
         if (filter["operator"] === ">=") {
-          if (value < parseInt(filter["value"])) {
+          if (value < parseFloat(filter["value"])) {
             return false;
           }
         } else if (filter["operator"] === "<=") {
-          if (value > parseInt(filter["value"])) {
+          if (value > parseFloat(filter["value"])) {
             return false;
           }
         }
